requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
bcrypt==4.1.2
boto3==1.34.95
botocore==1.34.95
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
ecdsa==0.19.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
jmespath==1.0.1
Mako==1.3.3
MarkupSafe==2.1.5
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
